fixed all the errors! it now can open in unity :D

+++ CMPM121Framework-main/Assets/Scripts/Levels/Enemy.cs
@@ -4,5 +4,5 @@ public class Enemy
     public int sprite;
     public int hp;
     public int speed;
-    public float damage; // switch to int?
+    public int damage; // switched from float to int
 }
